Fix bug where already-complete upgrade could go next
This is due to the fact that we are only checking the cache, and the
cache hasn't had a chance to update yet in some cases. This is fine, but
we should only look for upgrades that have not yet been processed yet,
not just upgrades that are not done yet (i.e. could still be marked as
in progress).